[feature]支持添加时判重：若与已加入表情包的MD5重复，则转到修改之前已加入的表情包

+++ app/src/main/java/com/skyd/rays/model/respository/AddRepository.kt
@@ -22,24 +22,45 @@ class AddRepository @Inject constructor(private val stickerDao: StickerDao) : Ba
         return flow {
             val tempFile = File(STICKER_DIR, "${Random.nextLong()}")
             uri.copyTo(tempFile)
-            stickerWithTags.sticker.stickerMd5 =
-                tempFile.md5() ?: error("can not calc sticker's md5!")
-            val uuid = stickerDao.addStickerWithTags(stickerWithTags)
-            tempFile.renameTo(File(STICKER_DIR, uuid))
-            tempFile.deleteRecursively()
-            emitBaseData(BaseData<String>().apply {
-                code = 0
-                data = uuid
-            })
+            val stickerMd5 = tempFile.md5() ?: error("can not calc sticker's md5!")
+            val containsByMd5 = stickerDao.containsByMd5(stickerMd5)
+            if (containsByMd5 != null &&
+                stickerDao.containsByUuid(stickerWithTags.sticker.uuid) == 0
+            ) {
+                tempFile.deleteRecursively()
+                emitBaseData(BaseData<String>().apply {
+                    code = -2
+                    msg = "Duplicate sticker!"
+                    data = containsByMd5
+                })
+            } else {
+                stickerWithTags.sticker.stickerMd5 = stickerMd5
+                val uuid = stickerDao.addStickerWithTags(stickerWithTags)
+                if (!tempFile.renameTo(File(STICKER_DIR, uuid))) {
+                    tempFile.deleteRecursively()
+                }
+                emitBaseData(BaseData<String>().apply {
+                    code = 0
+                    data = uuid
+                })
+            }
         }
     }
 
     suspend fun requestGetStickerWithTags(stickerUuid: String): Flow<BaseData<StickerWithTags>> {
         return flow {
-            emitBaseData(BaseData<StickerWithTags>().apply {
-                code = 0
-                data = stickerDao.getStickerWithTags(stickerUuid)
-            })
+            val stickerWithTags = stickerDao.getStickerWithTags(stickerUuid)
+            if (stickerWithTags == null) {
+                emitBaseData(BaseData<StickerWithTags>().apply {
+                    code = -1
+                    msg = "stickerWithTags is null"
+                })
+            } else {
+                emitBaseData(BaseData<StickerWithTags>().apply {
+                    code = 0
+                    data = stickerWithTags
+                })
+            }
         }
     }
 }
